Overhauling torchdrl to be more modular
Upated most of the agent files such that they are instead rl algorithms
instead of a standalone agent, now we have an agent that will generate
information from PlayEpisode and EvaluateEpisode.
I've moved the folder structure around too, and I've tested on cartpole,
so in theory it should be working.